+++ src/aulos/_core/utils/__init__.py
@@ -7,8 +7,9 @@ Provides generic helpers for intervals, positions, indexing, and type operations
 """
 
 from .dataclass import from_dict
+from .intervals import Intervals
+from .positions import Positions
 from .property import classproperty
-from .representation import Intervals, Positions
 from .sequence import index, rotated
 
 __all__ = [
